Continue fracturing into multiple files
